xml: manifest: Added the Ka-Ro flavoured build-environment setup script, incl. the inherent commit-id changes.
The Ka-Ro flavoured build-environment setup script retains data that sets
values for the required inclusion of the meta-karo-nxp layer into the Yocto
project.

The addition of the script to the layer itself inherently changed the top
commit-id.

Given that the script is essentially only a tweaked version of the NXP original
the original naming inside the meta layer has be be retained and the symlink
to for users consumption is flavoured according to the BSP flavouring.